update untwine for win

+++ external/untwine/bu/PointAccessor.hpp
@@ -48,10 +48,11 @@ public:
         m_fileInfos.push_back(&fi);
     }
 
-    Point operator[](int offset)
+    Point operator[](size_t offset)
     {
         for (FileInfo *fi : m_fileInfos)
-            if (offset >= fi->start() && offset < fi->start() + fi->numPoints())
+            if (offset >= (size_t)fi->start() &&
+                offset < (size_t)fi->start() + fi->numPoints())
                 return Point(fi->address() + ((offset - fi->start()) * m_b.pointSize));
         return Point();
     }
